rag-1/requirements.txt
openai
psycopg2-binary
python-dotenv
tiktoken